Add PAW module with AE-TDM support; fix POTCAR ordering and AUTDEBYE constant
- Add src/pawpot/: full PAW implementation including wavecar reader, POTCAR
  parser, G-vector generation, nonlocal projectors (nonlq), AE wavefunction
  reconstruction (aewfc), NormalCAR reader, and TDM via p-r relation with
  PAW one-centre correction (tdm_paw)
- Fix PawPotcar::from_str(): remove par_bridge() which caused non-deterministic
  POTCAR type ordering, making nablaij PAW corrections use wrong element data
- Fix AUTDEBYE constant in tdm_paw: 4.803242 -> 2.541746 (e·Bohr to Debye)
  to match VASP/VaspBandUnfolding convention
- Extend tdm command: add --poscar/--potcar flags to enable AE-TDM mode
- Add integration tests and reference data under tests/pawpot/
- Fix bench_tdm.rs: use wavecar.header.ngrid instead of hardcoded [0,0,0]

AE-TDM output now agrees with VaspBandUnfolding tdmplot to within 0.004 D
(<0.06%) across all tested MoS2 band pairs at k-point 1.

+++ src/vasp_parsers/soc.rs
@@ -52,7 +52,7 @@ where P: AsRef<Path> {
     let rec_l = f.read_i32::<LittleEndian>().unwrap();
     let nproj = rec_l / 16;
     f.seek(SeekFrom::Current(-4)).unwrap();
-    let mut cproj = na::Array3::<c64>::zeros((2, nbands as usize, nproj as usize));
+    let mut cproj = na::Array3::<c64>::zeros((2, nbands, nproj as usize));
     let mut buf = vec![0.0f64; nproj as usize * 2];
     for ispin in 0 .. 2 {
         for ikpt in 0 .. nkpoints {
@@ -62,7 +62,7 @@ where P: AsRef<Path> {
                 if ikpt + 1 == ikpoint {
                     f.read_f64_into::<LittleEndian>(&mut buf).unwrap();
                     for iproj in 0 .. nproj {
-                        cproj[(ispin as usize, iband as usize, iproj as usize)] = c64::new(buf[2 * iproj as usize], buf[2 * iproj as usize + 1]);
+                        cproj[(ispin as usize, iband, iproj as usize)] = c64::new(buf[2 * iproj as usize], buf[2 * iproj as usize + 1]);
                     }
                 } else {
                     f.seek(SeekFrom::Current(16 * nproj as i64)).unwrap();
